requirements.txt
dagster
dagit
pyyaml
requests
pymongo
gitpython